refactor:LangChain形式への変換をinfra層に移行
