+++ ai-api/app/infrastructure/vertex_chat_llm_client.py
@@ -1,9 +1,11 @@
 import os
 from langchain_google_vertexai import ChatVertexAI
+from app.domain.chat import Message
+from app.services.gateways.chat_llm_client import ChatLLMClient
 from app.infrastructure.config import vertex_config
 
 
-class VertexChatLLMClient:
+class VertexChatLLMClient(ChatLLMClient):
     """VertexAI クライアントクラス"""
     
     def __init__(self):
@@ -11,15 +13,21 @@ class VertexChatLLMClient:
          self._llm = ChatVertexAI(
              project=vertex_config.project_id,
              location=vertex_config.location,
-             model=vertex_config.model_name,           # 例: gemini-2.0-flash
+             model=vertex_config.model_name,
              temperature=float(os.getenv("TEMPERATURE", 0.7)),
              max_output_tokens=int(os.getenv("MAX_TOKENS", 1024)),
          )
     
-    async def chat(self, messages) -> str:
+    async def chat(self, messages: list[Message]) -> str:
         """
-        LangChain Message オブジェクト（System/Human/AI）を配列で受け取り、
-        非同期で 1 ターン返すだけ
+        Messageを配列で受け取り、メッセージを返す
         """
-        response = await self._llm.ainvoke(messages)
+        # LangChain形式に変換
+        langchain_messages = []
+        for message in messages:
+            langchain_messages.append({
+                "role": message.role,
+                "content": message.content
+            })
+        response = await self._llm.ainvoke(langchain_messages)
         return response.content
