+++ ai-api/app/services/gateways/chat_llm_client.py
@@ -1,0 +1,5 @@
+from typing import Protocol, List
+from app.domain.chat import Message
+
+class ChatLLMClient(Protocol):
+    async def chat(self, messages: List[Message]) -> str: ...
